Update Notifications page to delete/recover as they become active

+++ components/admin/notifications/NotificationDataTable.vue
@@ -55,7 +55,21 @@
               @search="onSearch"
               @edit-item="$refs.editor.open(null, $event)"
               @remove-item="remove"
-            />
+            >
+              <template v-slot:[`item.deletedAt`]="{ item }">
+                <span v-if="isNotificationActive(item)">
+                  ACTIVE
+                </span>
+                <span v-else>
+                  INACTIVE
+                </span>
+              </template>
+              <template v-slot:[`item.actions.prepend`]="{ item }">
+                <v-icon v-if="!isNotificationActive(item)" size="20" @click="restore(item)">
+                  mdi-history
+                </v-icon>
+              </template>
+            </pg-admin-data-table>
           </v-card-text>
         </v-card>
       </v-col>
@@ -95,9 +109,9 @@ export default {
           value: 'description'
         },
         {
-          text: 'Created',
+          text: 'Active',
           sortable: false,
-          value: 'createdAt'
+          value: 'deletedAt'
         },
         {
           text: 'Last Updated',
@@ -115,7 +129,7 @@ export default {
   },
 
   methods: {
-    ...mapActions('notifications', ['getNotifications', 'deleteNotification']),
+    ...mapActions('notifications', ['getNotifications', 'deleteNotification', 'restoreNotification']),
 
     async refresh (clear = false) {
       this.loading = true
@@ -125,22 +139,47 @@ export default {
       }
 
       try {
-        this.notifications = await this.getNotifications({ name: this.search })
+        this.notifications = await this.getNotifications({ name: this.search, deleted: true })
       } catch (e) {
       } finally {
         this.loading = false
       }
     },
 
-    remove ({ id, name }) {
+    remove ({ id, name, deletedAt }) {
+      if (this.isNotificationActive({ deletedAt })) {
+        this.$nuxt.$emit('open-prompt', {
+          message: `Are you sure you want to delete <b>${name}</b>?`,
+          action: async () => {
+            await this.deleteNotification(id)
+            await this.refresh()
+          }
+        })
+      } else {
+        this.$nuxt.$emit('open-prompt', {
+          message: 'Please, restore the notification first',
+          warning: 'This will allow you to edit and save changes',
+          actionText: 'Ok'
+        })
+      }
+    },
+
+    restore ({ id, name }) {
       this.$nuxt.$emit('open-prompt', {
-        title: 'Delete notification?',
-        message: `Are you sure you want to delete <b>${name}</b>?`,
+        message: `Are you sure you want to restore <b>${name}</b>?`,
+        warning: 'This will change the notification status to ACTIVE.',
+        actionText: 'Restore',
         action: async () => {
-          await this.deleteNotification(id)
+          await this.restoreNotification(id)
           await this.refresh()
         }
       })
+    },
+
+    /* goToProfile () {}, */
+
+    isNotificationActive (notification) {
+      return notification && notification.deletedAt === null
     }
   }
 }
